Moved SnapshotMap to components directory since it's no longer a class component
