+++ src/components/SnapshotMap/__tests__/SnapshotMap.test.js
